+++ src/app/view/exam-category/exam-category.component.ts
@@ -4,7 +4,7 @@ import { ButtonModule } from 'primeng/button';
 import { SidebarModule } from 'primeng/sidebar';
 import { AddExamCategoryComponent } from './add-exam-category/add-exam-category.component';
 import { ToastModule } from 'primeng/toast';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { ExamCategoryService } from '../../core/services/exam-category.service';
 import { TableModule } from 'primeng/table';
 import { PaginatorModule } from 'primeng/paginator';
@@ -12,36 +12,52 @@ import { TooltipModule } from 'primeng/tooltip';
 import { SubCategoryComponent } from "./sub-category/sub-category.component";
 import { TagModule } from 'primeng/tag';
 import { CategoryDetailComponent } from "./category-detail/category-detail.component";
+import { TabViewModule } from 'primeng/tabview';
+import { AddDigitalProductCategoryComponent } from './digital-product-category/add-digital-product-category/add-digital-product-category.component';
+import { ConfirmDialogModule } from 'primeng/confirmdialog';
 
 @Component({
   selector: 'app-exam-category',
   standalone: true,
-  imports: [CommonModule, ButtonModule, SidebarModule, AddExamCategoryComponent, ToastModule, TableModule, PaginatorModule, TooltipModule, SubCategoryComponent, TagModule, CategoryDetailComponent],
-  providers: [MessageService],
+  imports: [CommonModule, ButtonModule, SidebarModule, AddExamCategoryComponent, ConfirmDialogModule, ToastModule, TableModule, PaginatorModule, TooltipModule, SubCategoryComponent, TagModule, CategoryDetailComponent, TabViewModule, AddDigitalProductCategoryComponent ],
+  providers: [MessageService, ConfirmDialogModule],
   templateUrl: './exam-category.component.html',
   styleUrl: './exam-category.component.css'
 })
 export class ExamCategoryComponent {
   isSideBar: boolean = false;
+  isAddDigitalCategory : boolean = false;
+  editDigitalCategoryData : any;
   examCategoryList: any;
   isSubcategory: boolean = false;
   isViewDetails : boolean = false
 
   examCategoryId: any
-  subCategoryData : any
+  subCategoryData : any;
+  digitalproductDataList : any ;
 
   offset = 0;
   totalRecorde = 2;
   limit = 15;
 
-  constructor(private _messageService: MessageService, private _examCategory: ExamCategoryService) { }
+  constructor(private _messageService: MessageService, private _confirmationService: ConfirmationService, private _examCategory: ExamCategoryService) { }
 
   ngOnInit() {
     this.getExamCategoryAndSubCategoryList();
+    this.getDigitalProductCategory();
+
   }
 
   addExamCategory() {
     this.isSideBar = true
+  }
+
+  getDigitalProductCategory(){
+
+    this._examCategory.getDigitalProductCategory().subscribe(res => {
+      this.digitalproductDataList = res.data
+    })
+
   }
 
   getExamCategoryAndSubCategoryList() {
@@ -55,13 +71,15 @@ export class ExamCategoryComponent {
   closeSideBar(event: any) {
 
     if (event.status === true) {
-      this.isSideBar = false
+      this.isSideBar = false;
+      this.isAddDigitalCategory = false;
       this._messageService.add({ severity: 'success', detail: event.message });
     } else {
       this.isSideBar = false
     }
 
-    this.getExamCategoryAndSubCategoryList()
+    this.getExamCategoryAndSubCategoryList();
+    this.getDigitalProductCategory();
 
   }
 
@@ -102,20 +120,87 @@ export class ExamCategoryComponent {
 
   deleteCategory(id: any) {
 
-    const payload = {
-      "category_id": id
-    }
+    this._confirmationService.confirm({
+      header: '',
+      message: 'Are you sure. You want to delete category ?',
+      icon: 'null',
+      acceptButtonStyleClass: "danger-button text-base font-semibold",
+      rejectButtonStyleClass: "danger-border text-base button-text-danger bg-white font-semibold",
+      acceptLabel: "Yes",
+      acceptIcon: "none",
+      rejectLabel: "No",
+      rejectIcon: "none",
+      accept: () => {
 
-    this._examCategory.deleteCategory(payload).subscribe((res: any) => {
-      if (res.status === true) {
-        this._messageService.add({ severity: 'success', detail: res.message });
-        this.getExamCategoryAndSubCategoryList()
-      } else {
-        this._messageService.add({ severity: 'error', detail: res.message });
+        const payload = {
+          "category_id": id
+        }
+    
+        this._examCategory.deleteCategory(payload).subscribe((res: any) => {
+          if (res.status === true) {
+            this._messageService.add({ severity: 'success', detail: res.message });
+            this.getExamCategoryAndSubCategoryList()
+          } else {
+            this._messageService.add({ severity: 'error', detail: res.message });
+          }
+        })
+        
+      },
+      reject: () => {
+
       }
-    })
+
+    });
 
 
   }
+
+
+
+  // Digital Product Category 
+
+  addDigiCategory(){
+    this.isAddDigitalCategory = true
+  }
+
+  editDigiCategory(data : any){
+    this.isAddDigitalCategory = true;
+    this.editDigitalCategoryData = data
+  }
+
+  deleteDigiCategory(id : any){
+
+    this._confirmationService.confirm({
+      header: '',
+      message: 'Are you sure. You want to delete category ?',
+      icon: 'null',
+      acceptButtonStyleClass: "danger-button text-base font-semibold",
+      rejectButtonStyleClass: "danger-border text-base button-text-danger bg-white font-semibold",
+      acceptLabel: "Yes",
+      acceptIcon: "none",
+      rejectLabel: "No",
+      rejectIcon: "none",
+      accept: () => {
+    
+        this._examCategory.deleteDigitalCategory(id).subscribe((res: any) => {
+          if (res.status === true) {
+            this._messageService.add({ severity: 'success', detail: res.message });
+            this.getDigitalProductCategory();
+          } else {
+            this._messageService.add({ severity: 'error', detail: res.message });
+          }
+        })
+        
+      },
+      reject: () => {
+
+      }
+
+    });
+
+  }
+
+
+
 
 }
